connected backend and front end notifications
separated Notifications into its own component, and imported into TopBar
